requirements.txt
selenium 
requests 
Pyyaml
tenacity 